+++ VCalculateApi/Models/Aggregate.cs
@@ -3,6 +3,6 @@ namespace VCalculateApi.Models
     public class Aggregate
     {
         public float avg { get; set; }
-        public int sum { get; set; }
+        public float sum { get; set; }
     }
 }
